requirements.txt
Flask
flask_restful
mysql-connector-python-rf
yaml